chore: refactored a bit the reports page and its table
